+++ Changelog.md
@@ -1,5 +1,8 @@
 RestrictEvents Changelog
 ========================
+#### v1.0.1
+- Disabled `MacPro7,1` RAM & PCI Expansion Slot UIs
+- Disabled `MacBookAir` memory view restrictions
 
 #### v1.0.0
 - Initial release